v1.0.6 Changes
Added a command to toggle the plugin and re-worked the mechanics to use speed multipliers instead of fixed durations.
Added the ability to play a sound to the player when they wake up.
Documented the options in the config.yml, not that they'll stay there when Bukkit saves the file.